Add files via upload
Customer churn, also known as client attrition, is the phrase used to describe the likelihood of a current customer continuing to do business with the company. 
This parameter's probability factor is influenced by a variety of other factors in businesses such as banking, telecommunications, and a few others.
In today's harsh market circumstances, it's critical for businesses to keep track of customer churn and the numerous reasons why customers leave doing business with them.
So here this is a machine learning model which will predict the customers who are at risk.